qemu_plugin: Add a QEMU tracer plugin
This generates a shared object which can be provided to QEMU to
instrument all function calls and returns and to export a trace proto.

It can leverage the DWARF information of a target binary to do VMI
(Virtual Machine Introspection) and figure out the current process in
the guest. The DWARF parsing code is loosely adapted from the
Apache-licensed Google Bloaty. https://github.com/google/bloaty/

The trace contains information about the QEMU invokation so it can be
restarted for deterministic replay.