trace_processor: Add the ability to replay QEMU deterministic records
When instructed to "debug" a specific instruction number, the trace
processor will spawn a QEMU instance, regularly monitor its current
instruction number using the QEMU Monitor Port and report the status
over the IPC link. When the target instruction count is reached, it
spawns a gdb that attaches to the QEMU guest.